move job setup to jobs.py
